+++ ft_print_string.c
@@ -45,16 +45,16 @@ int			ft_print_low_str(va_list *ap, t_param *p)
 		str = ft_strdup("(null)");
 	if (p->prec > 0 && p->is_prec == 1)
 		str = ft_cut_str(str, p->prec);
-	if (p->width > ft_strlen(str))
+	if (p->width > (int)ft_strlen(str))
 	{
 		if (p->left == 1)
-			pos = ft_strlen(str);
+			pos = (int)ft_strlen(str);
 		else
 			pos = 0;
-		temp1 = ft_new_n_symb(p->width - ft_strlen(str), ft_c(p));
+		temp1 = ft_new_n_symb(p->width - (int)ft_strlen(str), ft_c(p));
 		str = ft_insert_with_free(str, temp1, pos);
 	}
-	len = ft_strlen(str);
+	len = (int)ft_strlen(str);
 	ft_putstr(str);
 	free(str);
 	return (len);
